重構 companyDetail 頁面
1. 將 companyDetail.{html,js} 之所有 templates 拆分為各別的檔案
2. 資料夾結構全面改用 panelFolders，廢棄原本的 rDisplayPanel 機制
3. 將一些相依於 template data context 的資料寫法改為相依於較為顯而易見的來源（如 paramCompany 等）
   同時去除部分 with 語法，以避免造成閱讀上的困擾
4. 部分重寫買賣單之顯示，現在可以同時顯示出自己的訂單在全體訂單的先後次序，並保留原有的列出自己訂單的功能
5. 將一些常用的全域性訂閱統一移至 util/globalSubscriptions.js 以方便管理
6. 將目前使用者的訂單與持股資訊改為全域性的訂閱寫法，避免切換頁面時不斷重複訂閱/取消而造成使用上的困擾
   （參照 #448）
