+++ client/arenaInfo/arenaInfo.js
@@ -68,15 +68,15 @@ Template.arenaInfo.events({
 Template.arenaInfoNav.helpers({
   arenaLinkAttrs(linkType) {
     const arenaId = FlowRouter.getParam('arenaId');
-    const currentArenaData = dbArena.findOne(arenaId);
+    const currentArena = dbArena.findOne(arenaId);
 
-    if (currentArenaData) {
+    if (currentArena) {
       switch (linkType) {
         case 'prev': {
           const navArenaData = dbArena.findOne(
             {
               beginDate: {
-                $lt: currentArenaData.beginDate
+                $lt: currentArena.beginDate
               }
             },
             {
@@ -104,7 +104,7 @@ Template.arenaInfoNav.helpers({
           const navArenaData = dbArena.findOne(
             {
               beginDate: {
-                $gt: currentArenaData.beginDate
+                $gt: currentArena.beginDate
               }
             },
             {
@@ -133,15 +133,15 @@ Template.arenaInfoNav.helpers({
   },
   arenaBegin() {
     const arenaId = FlowRouter.getParam('arenaId');
-    const currentArenaData = dbArena.findOne(arenaId);
+    const currentArena = dbArena.findOne(arenaId);
 
-    return currentArenaData ? currentArenaData.beginDate : null;
+    return currentArena ? currentArena.beginDate : null;
   },
   arenaEnd() {
     const arenaId = FlowRouter.getParam('arenaId');
-    const currentArenaData = dbArena.findOne(arenaId);
+    const currentArena = dbArena.findOne(arenaId);
 
-    return currentArenaData ? currentArenaData.endDate : null;
+    return currentArena ? currentArena.endDate : null;
   }
 });
