+++ client/company/companyDetailTable.js
@@ -1,0 +1,9 @@
+import { Template } from 'meteor/templating';
+
+import { paramCompany } from './helpers';
+
+Template.companyDetailTable.helpers({
+  company() {
+    return paramCompany();
+  }
+});
