+++ client/company/companyDirectorList.js
@@ -1,0 +1,69 @@
+import { ReactiveVar } from 'meteor/reactive-var';
+import { Template } from 'meteor/templating';
+import { Meteor } from 'meteor/meteor';
+
+import { dbDirectors } from '/db/dbDirectors';
+import { inheritedShowLoadingOnSubscribing } from '/client/layout/loading';
+import { alertDialog } from '/client/layout/alertDialog';
+import { getCurrentUserOwnedStockAmount, paramCompany, paramCompanyId } from './helpers';
+
+inheritedShowLoadingOnSubscribing(Template.companyDirectorList);
+
+Template.companyDirectorList.onCreated(function() {
+  this.offset = new ReactiveVar(0);
+
+  this.autorunWithIdleSupport(() => {
+    const companyId = paramCompanyId();
+    if (companyId) {
+      this.subscribe('companyDirector', companyId, this.offset.get());
+    }
+  });
+});
+
+Template.companyDirectorList.helpers({
+  company() {
+    return paramCompany();
+  },
+  directorList() {
+    const companyId = paramCompanyId();
+
+    return dbDirectors.find({ companyId }, { sort: { stocks: -1, createdAt: 1 } });
+  },
+  getPercentage(stocks) {
+    return Math.round(stocks / paramCompany().totalRelease * 10000) / 100;
+  },
+  getMessage(message) {
+    return message || '無';
+  },
+  paginationData() {
+    return {
+      useVariableForTotalCount: 'totalCountOfCompanyDirector',
+      dataNumberPerPage: 10,
+      offset: Template.instance().offset
+    };
+  },
+  getCurrentUserDirectorMessage() {
+    const userId = Meteor.user()._id;
+    const companyId = paramCompanyId();
+
+    return dbDirectors.findOne({ companyId, userId }).message;
+  },
+  isDirectorInVacation(userId) {
+    const user = Meteor.users.findOne(userId);
+
+    return user ? user.profile.isInVacation : false;
+  }
+});
+
+Template.companyDirectorList.events({
+  'submit form'(event, templateInstance) {
+    event.preventDefault();
+    const message = templateInstance.$('[name="message"]').val();
+    if (message.length > 100) {
+      alertDialog.alert('輸入訊息過長！');
+    }
+    else if (Meteor.user() && message.length) {
+      Meteor.customCall('directorMessage', paramCompanyId(), message);
+    }
+  }
+});
