+++ client/company/companyElectInfoSupporterListDialog.js
@@ -1,0 +1,8 @@
+import { Template } from 'meteor/templating';
+
+Template.companyElectInfoSupporterListDialog.events({
+  'click [data-action="dismiss"]'(event, templateInstance) {
+    event.preventDefault();
+    templateInstance.data.onDismiss();
+  }
+});
