+++ client/company/companyMiningMachine.html
