+++ client/company/companyMiningMachine.js
