+++ config.js
@@ -67,7 +67,8 @@ export const config = {
     userViolationCases: 10,
     userReportedViolationCases: 10,
     compoanyViolationCases: 10,
-    fscLogs: 30
+    fscLogs: 30,
+    companyOrders: 10
   },
   productFinalSaleTime: 86400000, // 產品最後出清時間 (ms)
   systemProductVotingReward: 4096, // 系統派發的推薦票回饋金
