+++ server/publications/company/currentUserDirectors.js
@@ -4,8 +4,8 @@ import { dbDirectors } from '/db/dbDirectors';
 import { limitSubscription } from '/server/imports/utils/rateLimit';
 import { debug } from '/server/imports/utils/debug';
 
-Meteor.publish('queryOwnStocks', function() {
-  debug.log('publish queryOwnStocks');
+Meteor.publish('currentUserDirectors', function() {
+  debug.log('publish currentUserDirectors');
   const userId = this.userId;
   if (userId) {
     return dbDirectors.find({ userId });
@@ -14,4 +14,4 @@ Meteor.publish('queryOwnStocks', function() {
   return [];
 });
 // 一分鐘最多20次
-limitSubscription('queryOwnStocks');
+limitSubscription('currentUserDirectors');
